Migrating to a Maven standard directory structure
Therefore the way to load test loop files and modules had to change in every tests calling `Loop.run()` and in the `loadAndCompile()` method (from `ModuleLoader`).
For both, the solution is to cache the relative path of each file by traversing the classpath at startup and load the file on demand by retrieving the path of a module from the cache.